refactor(Locations): Remove museum locations data
We prefer not to publish an artwork's current location in the API data because locations can change frequently or suddenly. (For example, if a gallery ever has to be closed for maintenance, then every object therein becomes "not on view," then they all change again when the gallery reopens) In the same way, users of our data should prefer to omit the location data since it can quickly become out-of-date and require more updates, whether via API or a scrape. Our collection website remains our preferred/only source for the location data.

As an aside, the names of our locations, the structure of that data, etc., are up for review and likely to change. The "locations" data might not be published in API v2. Removing it for now. see also Issue #85

BREAKING CHANGE: